fix(srf_parser): missed a spot!

+++ source_modelling/srf_parser/src/lib.rs
@@ -3,7 +3,6 @@ use numpy::PyArray1;
 use pyo3::exceptions::PyOSError;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
-use pyo3::types::PyDict;
 use pyo3::wrap_pyfunction;
 
 use std::fs::File;
@@ -119,7 +118,7 @@ fn read_srf_points(
             slipt1.data.push(slip);
         }
     }
-    slipt1.row_ptr.push(slipt1.data.len());
+    slipt1.row_ptr.push(slipt1.data.len() as u32);
     Ok((metadata, slipt1))
 }
